requirements.txt
pyrealsense2>=2.50
opencv-python>=4.8
numpy>=1.24
PyYAML>=6.0
